refactor: Moving lts namespace
