+++ packages/lts/src/main/java/com/bt/scrumble/lts/ScrumbleLtsApi.java
@@ -1,4 +1,4 @@
-package com.bt.scrumbleLts;
+package com.bt.scrumble.lts;
 
 import com.blade.Blade;
