+++ packages/lts/src/main/java/com/bt/scrumble/lts/api/ProjectApi.java
@@ -1,9 +1,9 @@
-package com.bt.scrumbleLts.api;
+package com.bt.scrumble.lts.api;
 
 import com.blade.mvc.annotation.GetRoute;
 import com.blade.mvc.annotation.JSON;
 import com.blade.mvc.annotation.Path;
-import com.bt.scrumbleLts.dto.Project;
+import com.bt.scrumble.lts.dto.Project;
 
 @Path
 public class ProjectApi {
